Deep_Python/Plotly and dash practice/requirements.txt
dash==2.14.1
numpy==1.23.5
pandas==1.5.3
plotly==5.18.0
python_dateutil==2.8.2
wget==3.2
